# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Starter API",
    "contact": {}
  },
  "host": "localhost:8080.",
  "paths": {
...
    "/api/v1/booking": {
      "post": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "summary": "Бронирование креативной площадки",
        "parameters": [
          {
            "description": "Параметры запроса",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/requests.createBookingRequestData"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/model.ResponseSuccess"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/responses.createBookingResponseData"
                    }
                  }
                }
              ]
            }
          },
          "403": {
            "description": "Коды ошибок: [1100]",
            "schema": {
              "$ref": "#/definitions/model.ResponseError"
            }
          }
        }
      }
    },
    "/api/v1/bookings": {
      "get": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "summary": "Возвращает список бронирований",
        "parameters": [
          {
            "type": "string",
            "description": "Authentication header",
            "name": "Authorization",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/model.ResponseSuccess"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/responses.getBookingsResponseData"
                    }
                  }
                }
              ]
            }
          },
          "403": {
            "description": "Коды ошибок: [1100]",
            "schema": {
              "$ref": "#/definitions/model.ResponseError"
            }
          }
        }
      }
    },
    "/api/v1/bookings/:id": {
      "get": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "summary": "Возвращает бронирование по id",
        "parameters": [
          {
            "type": "string",
            "description": "Authentication header",
            "name": "Authorization",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/model.ResponseSuccess"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/responses.getBookingResponseData"
                    }
                  }
                }
              ]
            }
          },
          "403": {
            "description": "Коды ошибок: [1100]",
            "schema": {
              "$ref": "#/definitions/model.ResponseError"
            }
          }
        }
      },
      "delete": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "summary": "Удаление информации о бронировании",
        "parameters": [
          {
            "type": "string",
            "description": "Id бронирования",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/model.ResponseSuccess"
            }
          },
          "403": {
            "description": "Коды ошибок: [1100]",
            "schema": {
              "$ref": "#/definitions/model.ResponseError"
            }
          }
        }
      },
      "patch": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "summary": "Изменение дат бронирования креативной площадки",
        "parameters": [
          {
            "description": "Параметры запроса",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/requests.patchBookingRequestData"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/model.ResponseSuccess"
            }
          },
          "403": {
            "description": "Коды ошибок: [1100]",
            "schema": {
              "$ref": "#/definitions/model.ResponseError"
            }
          }
        }
      }
    },
    "/api/v1/calendar/convert": {
      "post": {
        "consumes": ["multipart/form-data"],
        "produces": ["application/json"],
        "summary": "Возвращает события календаря",
        "parameters": [
          {
            "type": "string",
            "description": "Authentication header",
            "name": "Authorization",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/model.ResponseSuccess"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/responses.convertCalendarResponseData"
                    }
                  }
                }
              ]
            }
          },
          "403": {
            "description": "Коды ошибок: [1100]",
            "schema": {
              "$ref": "#/definitions/model.ResponseError"
            }
          }
        }
      }
    },
    "/api/v1/calendar/convertFromLink": {
      "post": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "summary": "Возвращает события календаря",
        "parameters": [
          {
            "type": "string",
            "description": "Authentication header",
            "name": "Authorization",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/model.ResponseSuccess"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/responses.convertCalendarFromLinkResponseData"
                    }
                  }
                }
              ]
            }
          },
          "403": {
            "description": "Коды ошибок: [1100]",
            "schema": {
              "$ref": "#/definitions/model.ResponseError"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "model.ResponseError": {
      "type": "object",
      "properties": {
        "error": {
          "$ref": "#/definitions/model.ResponseErrorField"
        },
        "success": {
          "type": "boolean",
          "default": false
        }
      }
    },
    "model.ResponseErrorField": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "details": {
          "type": "string"
        }
      }
    },
    "model.ResponseSuccess": {
      "type": "object",
      "properties": {
        "data": {},
        "success": {
          "type": "boolean",
          "default": true
        }
      }
    },
    "requests.createBookingRequestCalendarEvent": {
      "type": "object",
      "required": ["date"],
      "properties": {
        "date": {
          "type": "string"
        }
      }
    },
    "requests.createBookingRequestData": {
      "type": "object",
      "required": ["calendarEvents", "creativeSpaceId"],
      "properties": {
        "calendarEvents": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/requests.createBookingRequestCalendarEvent"
          }
        },
        "creativeSpaceId": {
          "type": "integer"
        }
      }
    },
...
    "requests.patchBookingRequestCalendarEvent": {
      "type": "object",
      "required": ["date"],
      "properties": {
        "date": {
          "type": "string"
        }
      }
    },
    "requests.patchBookingRequestData": {
      "type": "object",
      "properties": {
        "calendarEvents": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/requests.patchBookingRequestCalendarEvent"
          }
        },
        "id": {
          "type": "integer"
        },
        "status": {
          "type": "string"
        }
      }
    },
...
    "responses.convertCalendarFromLinkResponseData": {
      "type": "object",
      "properties": {
        "events": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/responses.convertCaledarFromLinkResponseEvent"
          }
        }
      }
    },
    "responses.convertCaledarFromLinkResponseEvent": {
      "type": "object",
      "properties": {
        "date": {
          "type": "string"
        }
      }
    },
    "responses.convertCalendarResponseData": {
      "type": "object",
      "properties": {
        "events": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/responses.convertCaledarResponseEvent"
          }
        }
      }
    },
    "responses.convertCaledarResponseEvent": {
      "type": "object",
      "properties": {
        "date": {
          "type": "string"
        }
      }
    },
    "responses.createBookingResponseBooking": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        }
      }
    },
    "responses.createBookingResponseData": {
      "type": "object",
      "properties": {
        "booking": {
          "$ref": "#/definitions/responses.createBookingResponseBooking"
        }
      }
    },
...
    "responses.getBookingResponseBooking": {
      "type": "object",
      "properties": {
        "calendarEvents": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/responses.getBookingResponseCalendarEvent"
          }
        },
        "creativeSpace": {
          "$ref": "#/definitions/responses.getBookingResponseCreativeSpace"
        },
        "fullPrice": {
          "type": "integer"
        },
        "id": {
          "type": "integer"
        },
        "landlordInfo": {
          "$ref": "#/definitions/responses.getBookingResponseLandlordInfo"
        },
        "scoreId": {
          "type": "integer"
        },
        "status": {
          "type": "string"
        },
        "tenantInfo": {
          "$ref": "#/definitions/responses.getBookingResponseTenantInfo"
        }
      }
    },
    "responses.getBookingResponseCalendarEvent": {
      "type": "object",
      "properties": {
        "date": {
          "type": "string"
        }
      }
    },
    "responses.getBookingResponseCreativeSpace": {
      "type": "object",
      "properties": {
        "address": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "photos": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "title": {
          "type": "string"
        }
      }
    },
    "responses.getBookingResponseData": {
      "type": "object",
      "properties": {
        "booking": {
          "$ref": "#/definitions/responses.getBookingResponseBooking"
        }
      }
    },
    "responses.getBookingResponseLandlordInfo": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "inn": {
          "type": "string"
        },
        "legalEntityName": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "patronymic": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "role": {
          "type": "string"
        },
        "surname": {
          "type": "string"
        }
      }
    },
    "responses.getBookingResponseTenantInfo": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "patronymic": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "role": {
          "type": "string"
        },
        "surname": {
          "type": "string"
        }
      }
    },
    "responses.getBookingsResponseBooking": {
      "type": "object",
      "properties": {
        "calendarEvents": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/responses.getBookingsResponseCalendarEvent"
          }
        },
        "creativeSpace": {
          "$ref": "#/definitions/responses.getBookingsResponseCreativeSpace"
        },
        "fullPrice": {
          "type": "integer"
        },
        "id": {
          "type": "integer"
        },
        "landlordInfo": {
          "$ref": "#/definitions/responses.getBookingsResponseLandlordInfo"
        },
        "scoreId": {
          "type": "integer"
        },
        "status": {
          "type": "string"
        },
        "tenantInfo": {
          "$ref": "#/definitions/responses.getBookingsResponseTenantInfo"
        }
      }
    },
    "responses.getBookingsResponseCalendarEvent": {
      "type": "object",
      "properties": {
        "date": {
          "type": "string"
        }
      }
    },
    "responses.getBookingsResponseCreativeSpace": {
      "type": "object",
      "properties": {
        "address": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "photos": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "title": {
          "type": "string"
        }
      }
    },
    "responses.getBookingsResponseData": {
      "type": "object",
      "properties": {
        "bookings": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/responses.getBookingsResponseBooking"
          }
        }
      }
    },
    "responses.getBookingsResponseLandlordInfo": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "inn": {
          "type": "string"
        },
        "legalEntityName": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "patronymic": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "role": {
          "type": "string"
        },
        "surname": {
          "type": "string"
        }
      }
    },
    "responses.getBookingsResponseTenantInfo": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "patronymic": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "role": {
          "type": "string"
        },
        "surname": {
          "type": "string"
        }
      }
    },
...
  }
}